pyproject.toml
[project]
name = "sea-retail-media-ssot"
version = "0.1.0"
description = "Supermetrics → BigQuery ELT for SEA retail media (Shopee + TikTok Shop anchor)"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-secret-manager>=2.20.0",
    "google-cloud-storage>=2.18.0",
    "requests>=2.32.0",
    "tenacity>=8.5.0",
    "pydantic>=2.8.0",
    "pyyaml>=6.0.2",
    "click>=8.1.7",
    "python-dateutil>=2.9.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]
sim = [
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "scipy>=1.14.0",
]

[project.scripts]
ssot = "ssot.main:cli"
gen-data = "ssot.sim.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ssot"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
